+++ extensions/youtube-downloader/CHANGELOG.md
@@ -1,5 +1,10 @@
 # YouTube Downloader Changelog
 
+## [Improve Error Message] - {PR_MERGE_DATE}
+
+- Improve error message
+- Replace `execa` with `nano-spawn`
+
 ## [Improve URL Validator] - 2025-01-23
 
 - Improve `isYouTubeURL` function